# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Träwelling API",
        "description": "Träwelling user API description. This is an incomplete documentation with still many errors. The API is currently not yet stable. Endpoints are still being restructured. Both the URL and the request or body can be changed. Breaking changes will be announced on GitHub: https://github.com/Traewelling/traewelling/blob/develop/API_CHANGELOG.md",
        "contact": {
            "email": "support@traewelling.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0 - alpha"
    },
    "servers": [
        {
            "url": "https://traewelling.de/api/v1",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "This instance"
        }
    ],
    "paths": {
...
        "/user/{id}/follow": {
            "post": {
                "tags": [
                    "User/Follow"
                ],
                "summary": "Follow a user",
                "operationId": "createFollow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User-ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1337
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "Already following"
                    },
                    "403": {
                        "description": "User is blocked"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-follows"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User/Follow"
                ],
                "summary": "Unfollow a user",
                "operationId": "destroyFollow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User-ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1337
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Already following"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-follows"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
        "/user/self/followers": {
            "get": {
                "tags": [
                    "User/Follow",
                    "Settings"
                ],
                "summary": "List all followers",
                "operationId": "getFollowers",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "Already following"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-settings-followers"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
        "/user/self/follow-requests": {
            "get": {
                "tags": [
                    "User/Follow",
                    "Settings"
                ],
                "summary": "List all followers",
                "operationId": "getFollowRequests",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-settings-followers"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
        "/user/self/followings": {
            "get": {
                "tags": [
                    "User/Follow",
                    "Settings"
                ],
                "summary": "List all users the current user is following",
                "operationId": "getFollowings",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-settings-followers"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
        "/user/self/followers/{userId}": {
            "delete": {
                "tags": [
                    "User/Follow"
                ],
                "summary": "Remove a follower",
                "operationId": "removeFollower",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User-ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1337
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Follow not found"
                    },
                    "500": {
                        "description": "Unknown error"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-followers"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
        "/user/self/follow-requests/{userId}": {
            "put": {
                "tags": [
                    "User/Follow"
                ],
                "summary": "Accept a follow request",
                "operationId": "acceptFollowRequest",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User-ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1337
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-followers"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User/Follow"
                ],
                "summary": "Reject a follow request",
                "operationId": "rejectFollowRequest",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User-ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1337
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-followers"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int",
                        "example": 1
                    },
                    "displayName": {
                        "title": "displayName",
                        "description": "Display name of the user",
                        "example": "Gertrud"
                    },
                    "username": {
                        "title": "username",
                        "description": "username of user",
                        "type": "string",
                        "example": "Gertrud123"
                    },
                    "profilePicture": {
                        "title": "profilePicture",
                        "description": "URL of the profile picture of the user",
                        "type": "integer",
                        "example": "https://traewelling.de/@Gertrud123/picture"
                    },
                    "trainDistance": {
                        "title": "trainDistance",
                        "description": "distance travelled by train in meters",
                        "type": "integer",
                        "format": "int",
                        "example": 12345
                    },
                    "trainDuration": {
                        "title": "trainDuration",
                        "description": "duration travelled by train in minutes",
                        "type": "integer",
                        "format": "int",
                        "example": 6
                    },
                    "points": {
                        "title": "points",
                        "description": "Current points of the last 7 days",
                        "type": "integer",
                        "format": "int",
                        "example": 300
                    },
                    "mastodonUrl": {
                        "title": "mastodonUrl",
                        "description": "URL to the Mastodon profile of the user",
                        "type": "string",
                        "example": "https://chaos.social/@traewelling",
                        "nullable": true
                    },
                    "privateProfile": {
                        "title": "privateProfile",
                        "description": "is this profile set to private?",
                        "type": "boolean",
                        "example": false
                    },
                    "likes_enabled": {
                        "title": "likes_enabled",
                        "description": "Does this profile allow likes? Only offer the UI to like any status if this setting is set to true. If set to false, the likes API will return 403.",
                        "type": "boolean",
                        "example": true
                    },
                    "userInvisibleToMe": {
                        "title": "userInvisibleToMe",
                        "description": "Can the currently authenticated user see the statuses of this user?",
                        "type": "boolean",
                        "example": false
                    },
                    "muted": {
                        "title": "muted",
                        "description": "Is this user muted by the currently authenticated user?",
                        "type": "boolean",
                        "example": false
                    },
                    "following": {
                        "title": "following",
                        "description": "Does the currently authenticated user follow this user?",
                        "type": "boolean",
                        "example": false
                    },
                    "followPending": {
                        "title": "followPending",
                        "description": "Is there a currently pending follow request?",
                        "type": "boolean",
                        "example": false
                    },
                    "followedBy": {
                        "title": "followedBy",
                        "description": "Is the user following you?",
                        "type": "boolean",
                        "example": false
                    },
                    "preventIndex": {
                        "title": "preventIndex",
                        "description": "Did the user choose to prevent search engines from indexing their profile?",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 authorizationCode Flow. Get your token from https://traewelling.de/settings/applications. Set the redirect URL to https://traewelling.de/api/oauth2-callback to use it here. See also https://laravel.com/docs/9.x/passport",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/auth/refresh",
                        "scopes": {
                            "read-statuses": "see all statuses",
                            "read-notifications": "see your notifications",
                            "read-statistics": "see your statistics",
                            "read-search": "search in Träwelling",
                            "write-statuses": "create, edit, delete statuses",
                            "write-likes": "create and remove likes",
                            "write-notifications": "mark notifications as read, clear notifications",
                            "write-exports": "request data exports",
                            "write-follows": "follow and unfollow users",
                            "write-followers": "accept follow requests and remove followers",
                            "write-blocks": "block and unblock users, mute and unmute users",
                            "write-event-suggestions": "suggest events in your name",
                            "write-support-tickets": "create support tickets in your name",
                            "read-settings": "see your settings, email, etc.",
                            "write-settings-profile": "edit your profile",
                            "read-settings-profile": "see your profile data, e.g. email",
                            "write-settings-mail": "edit your email",
                            "write-settings-profile-picture": "edit your profile picture",
                            "write-settings-privacy": "change your privacy settings",
                            "read-settings-followers": "show follow-requests and followers",
                            "write-settings-calendar": "create and delete new calendar-tokens",
                            "extra-write-password": "change your password",
                            "extra-terminate-sessions": "log you out of other sessions and apps",
                            "extra-delete": "delete your account"
                        }
                    }
                }
            },
            "token": {
...
    },
...
}